The `measure [--real-time]` operator emits metrics about the input into
the pipeline. The operator handles both events and bytes as input, and
emits the metrics as events as described by these schemas, respectively:

```
type vast.metrics.events = record  {
  timestamp: time,
  schema: string,
  schema_id: string,
  events: uint64,
}

type vast.metrics.bytes = record  {
  timestamp: time,
  bytes: uint64,
}
```

The `--real-time` option causes the operator to instantly emit the
metrics at the cost of throughput.

+++ libvast/include/vast/defaults.hpp
@@ -31,7 +31,7 @@ inline constexpr size_t max_recursion = 100;
 namespace import {
 
 /// Maximum size for sources that generate table slices.
-inline constexpr size_t table_slice_size = 65'536; // 64 Ki
+inline constexpr uint64_t table_slice_size = 65'536; // 64 Ki
 
 /// The default table slice type when arrow is available.
 inline constexpr auto table_slice_type = table_slice_encoding::arrow;
